+++ demo/module2.py
@@ -1,0 +1,7 @@
+from module1 import nice_color
+
+print("I am", __name__)
+print(nice_color)
+
+if __name__ == "__main__":
+    print("module 2 is main")