+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `fw_version` attribute to `pybricksdev.connections.pybricks.PybricksHub`.
+- Experimental support for multi-file projects.
 
 ### Fixed
 - Fixed running programs on hubs with firmware with MPY ABI v5.