+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-alpha.30] - 2022-08-26
+
 ### Added
 - Added `fw_version` attribute to `pybricksdev.connections.pybricks.PybricksHub`.
 - Experimental support for multi-file projects.
@@ -267,8 +269,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.29..HEAD
-[1.0.0-alpha.28]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.28...v1.0.0-alpha.29
+[Unreleased]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.30..HEAD
+[1.0.0-alpha.30]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.29...v1.0.0-alpha.30
+[1.0.0-alpha.29]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.28...v1.0.0-alpha.29
 [1.0.0-alpha.28]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.27...v1.0.0-alpha.28
 [1.0.0-alpha.27]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.26...v1.0.0-alpha.27
 [1.0.0-alpha.26]: https://github.com/pybricks/pybricksdev/compare/v1.0.0-alpha.25...v1.0.0-alpha.26